LPD-45896 Isolating the new tests to avoid breaking the old ones with the ff enabled
